requirements.txt
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.10.2
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
bcrypt==4.0.1
certifi==2023.7.22
cloudinary==1.32.0
cloudpathlib==0.16.0
cryptography==41.0.5
email-validator==1.3.1
fastapi==0.95.0
fastapi-limiter==0.1.5
fastapi-mail==1.2.7
Jinja2==3.1.2
opencv-python==4.7.0.72
psycopg2-binary
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
redis==4.5.4
SQLAlchemy==2.0.7
starlette==0.26.1
uvicorn==0.21.1
urllib3==1.26.18
validators==0.22.0
libgravatar==1.0.4
passlib==1.7.4
python-multipart==0.0.6
frozenlist==1.4.0
python-dateutil==2.8.2
python-iso639==2023.6.15
colorama==0.4.6
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
watchfiles==0.21.0
websockets==12.0

